+++ CMakeLists.txt
@@ -159,6 +159,7 @@ install(FILES src/Interface/hiopInterface.hpp
 ##########################################################
 if (HIOP_WITH_MAKETEST)
   enable_testing()
+  add_test(NAME VectorTest        COMMAND $<TARGET_FILE:testVector> -selfcheck)
   add_test(NAME NlpDenseCons1_5H  COMMAND $<TARGET_FILE:nlpDenseCons_ex1.exe>   500 1.0 -selfcheck)
   add_test(NAME NlpDenseCons1_5K  COMMAND $<TARGET_FILE:nlpDenseCons_ex1.exe>  5000 1.0 -selfcheck)
   add_test(NAME NlpDenseCons1_50K COMMAND $<TARGET_FILE:nlpDenseCons_ex1.exe> 50000 1.0 -selfcheck)
