+++ tests/CMakeLists.txt
@@ -1,4 +1,4 @@
 
 
-add_executable(testVector testVector.cpp vectorTestsPar.cpp)
+add_executable(testVector testVector.cpp LinAlg/vectorTestsPar.cpp)
 target_link_libraries(testVector hiop ${HIOP_MATH_LIBRARIES})
