+++ tests/LinAlg/vectorTests.hpp
@@ -14,6 +14,26 @@ public:
         return x.get_size() == N ? 0 : 1;
     }
 
+    int testSetToConstant(hiop::hiopVector& x)
+    {
+        int N = x.get_size();
+
+        for(int i=0; i<N; ++i)
+        {
+            setElement(&x, i, 0.0);
+        }
+
+        x.setToConstant(1.0);
+
+        for(int i=0; i<N; ++i)
+        {
+            if (getElement(&x, i) != 1.0)
+                return 1;
+        }
+
+        return 0;
+    }
+
 protected:
     virtual void   setElement(hiop::hiopVector* x, int i, double val) = 0;
     virtual double getElement(const hiop::hiopVector* x, int i) = 0;
